Rename to vllm-ascend
