+++ vllm_ascend/ops/__init__.py
@@ -1,0 +1,1 @@
+import vllm_ascend.ops.layernorm # noqa
